+++ linear_solver.h
@@ -3,6 +3,6 @@
 
 #include "tridiagonal_matrix.h"
 
-int linear_system_solver(struct tridiagonal_matrix *m, double b[], double *result);
-int tridiag_system_solver(struct tridiagonal_matrix *m, double b[], double *result);
+double linear_system_solver(struct tridiagonal_matrix *m, double b[], double *result);
+double tridiag_system_solver(struct tridiagonal_matrix *m, double b[], double *result);
 #endif // !__LINEAR_SOLVER
